requirements.txt
google-generativeai==0.8.3
streamlit==1.29.0
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain-core==0.3.29
pydantic>=2.0.0
